Add functions to produce layered structure
Produce a list of generational layers for the use of the tree drawing
routine. The current structure took some trial and error, and may
still change further.

The scheme as it stands is:

- And outer list with an index for each generation (there's no current
  accommodation for people managing to span generational layers)
- A dictionary for each layer containing the following keys:
    - 'groups': a subdictionary of parents_ids against the respective
                groups of children
    - 'edges': a subdictionary of couple IDs (i.e. the children's
               parent_id) against tuples of the parents

The idea behind labelling the latter 'edges' rather than e.g. 'couples'
is that decisions like left-to-right ordering (particularly when there
are children from several marriages, etc.) will be made by the Family
or Person class, rather than by the tree-drawing script.