First pass at family tree diagram drawing
It's a horrible monolithic difficult-to-comprehend routine in sore need
of refactoring. It's a bit buggy at the best of times, and breaks down
completely for larger trees. It doesn't properly support cases of only
one known parent, or cases of multiple relationships. The formatting
and styling isn't finalized.

But it does kind of work, finally.