Use lists rather than sets for layer groups
As handy as sets are for ensuring uniqueness, retaining the sequence
in which people were added helps a lot with keeping the diagram under
control, as the tree is explored in a sensible order. Things still
break down for large diagrams, but at smaller scales it works a lot
better.