Fix node order and kinked edges in tree diagram
The initial kinked edges were caused by the invisible edges between
the first node of each layer in order to maintain layer hierarchy. This
was fixed by the introduction of the invisible achor node hierarchy in
a previous commit.

The order of invisible header nodes (i.e. the nodes used at branch
points) was corrected by creating invisible edges directly from the
left or right parent to that node, depending on which half of the
trunk they fall (the visible edge is drawn from the relationship node
between the parents).

Unfortunately the above solution caused edges to kink again, as the
invisible edge would sometimes take up the side edge, forcing the
visible edge to route to the top instead. This was resolved by adding
further invisible nodes to either side of the trunk edge.