Treat zero-precision DOB and DOD as absent
Zero-precision dates are a handy workaround for when e.g. birth order
is known, but not the actual years. In these instances the `.age` and
`.years` cached properties will return `None`, which elegantly avoids
printing any erroneous information, while allowing the dummy DOB/DOD to
be used for sorting purposes.