Refactor config as singleton class
This new `Config` class mainly functions as a thin semantic wrapper
for the relevant `ConfigParser` calls. It will read and cache the config
file the first time the user attempts to access a section; all section
data is retrieved from this cache. This is more for the sake of avoiding
unexpected behaviour arising from inconsistent configuration values
rather than any performance benefit.